requirements.txt
# Web server
Flask==3.0.2
gunicorn==21.2.0

# Task workers
redis==5.0.1
celery==5.3.6

# Dev stuff
flask-debugtoolbar==0.14.1

# App requirements
requests==2.31.0
selenium==4.17.2
PyYAML==6.0.1
PyNaCl==1.5.0